requirements.txt
msgpack~=1.0.2
Pillow~=8.1.0
pytz~=2021.1
aiohttp~=3.7.3